backend/requirements.txt
asgiref==3.9.1
Django==5.1.1
django-cors-headers==4.4.0
django-environ==0.11.2
djangorestframework==3.15.2
psycopg2-binary==2.9.9
sqlparse==0.5.3
whitenoise==6.7.0
django-tenants>=3.6
djangorestframework-simplejwt>=5.3
drf-spectacular>=0.27
